extracted services for queue and session to another module
