+++ REST/pom.xml
@@ -26,6 +26,14 @@
 
     <dependencies>
         <dependency>
+            <groupId>pl.jakubowskiprzemyslaw.tajgertim</groupId>
+            <artifactId>QueueAndSessionService</artifactId>
+            <version>1.0</version>
+            <scope>system</scope>
+            <systemPath>${basedir}/../QueueAndSessionService/target/QueueAndSessionService-1.0-jar-with-dependencies.jar</systemPath>
+        </dependency>
+
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
@@ -49,6 +57,12 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-amqp</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>QueueAndSessionService</artifactId>
+            <version>2.0.3.RELEASE</version>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
