+++ REST/src/main/java/pl/jakubowskiprzemyslaw/REST/configuration/RabbitMQConfiguration.java
@@ -1,6 +1,7 @@
 package pl.jakubowskiprzemyslaw.REST.configuration;
 
 import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -27,5 +28,4 @@ public class RabbitMQConfiguration {
   public Queue getPlayerConfigQueue() {
     return new Queue(playerConfigQueueName, durable, exclusive, autoDelete);
   }
-
 }
